# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/students/{id}/frequencia": {
      "get": {
        "operationId": "StudentsController_frequencia",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dias",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrequenciaDoAlunoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/api/v1/students": {
      "get": {
        "operationId": "StudentsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "vinculo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pendente",
                "aprovado",
                "recusado"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoPaginadoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoComSenhaTemporariaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/api/v1/students/{id}/senha-temporaria": {
      "post": {
        "operationId": "StudentsController_regenerarSenhaTemporaria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoComSenhaTemporariaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/api/v1/students/{id}/aprovar": {
      "post": {
        "operationId": "StudentsController_aprovar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/api/v1/students/{id}/recusar": {
      "post": {
        "operationId": "StudentsController_recusar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "operationId": "StudentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "patch": {
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
...
    "/api/v1/teachers/{id}/foto": {
      "put": {
        "operationId": "TeacherPhotoController_substituir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "arquivo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FotoDeProfessorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "teachers"
        ]
      },
      "delete": {
        "operationId": "TeacherPhotoController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FotoDeProfessorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "teachers"
        ]
      }
    },
...
    "/api/v1/classes/{id}/students/{alunoId}": {
      "post": {
        "operationId": "ClassesController_allocateStudent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatriculaEmTurmaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classes"
        ]
      },
      "delete": {
        "operationId": "ClassesController_removeStudent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "classes"
        ]
      }
    },
...
  },
  "info": {
    "title": "PlayCK API",
    "description": "Contrato da API do back — fonte para geração de tipos nos 3 frontends (ADR-001)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FaltasSeguidasComposicaoResponseDto": {
        "type": "object",
        "properties": {
          "ausente": {
            "type": "number",
            "example": 2
          },
          "justificado": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "ausente",
          "justificado"
        ]
      },
      "AgregadoDeFrequenciaResponseDto": {
        "type": "object",
        "properties": {
          "frequenciaPct": {
            "type": "number",
            "nullable": true,
            "example": 83
          },
          "confianca": {
            "type": "string",
            "enum": [
              "alta",
              "baixa"
            ]
          },
          "base": {
            "type": "number",
            "example": 12
          },
          "presente": {
            "type": "number",
            "example": 10
          },
          "ausente": {
            "type": "number",
            "example": 1
          },
          "justificado": {
            "type": "number",
            "example": 1
          },
          "faltasSeguidas": {
            "type": "number",
            "example": 3
          },
          "faltasSeguidasComposicao": {
            "$ref": "#/components/schemas/FaltasSeguidasComposicaoResponseDto"
          }
        },
        "required": [
          "frequenciaPct",
          "confianca",
          "base",
          "presente",
          "ausente",
          "justificado",
          "faltasSeguidas",
          "faltasSeguidasComposicao"
        ]
      },
      "CoberturaResponseDto": {
        "type": "object",
        "properties": {
          "aconteceram": {
            "type": "number",
            "example": 12
          },
          "lancadas": {
            "type": "number",
            "example": 9
          },
          "completas": {
            "type": "number",
            "example": 7
          },
          "pctCompletas": {
            "type": "number",
            "nullable": true,
            "example": 58
          },
          "confianca": {
            "type": "string",
            "enum": [
              "alta",
              "baixa"
            ]
          },
          "aviso": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "aconteceram",
          "lancadas",
          "completas",
          "pctCompletas",
          "confianca",
          "aviso"
        ]
      },
      "FrequenciaPorTurmaResponseDto": {
        "type": "object",
        "properties": {
          "frequenciaPct": {
            "type": "number",
            "nullable": true,
            "example": 83
          },
          "confianca": {
            "type": "string",
            "enum": [
              "alta",
              "baixa"
            ]
          },
          "base": {
            "type": "number",
            "example": 12
          },
          "presente": {
            "type": "number",
            "example": 10
          },
          "ausente": {
            "type": "number",
            "example": 1
          },
          "justificado": {
            "type": "number",
            "example": 1
          },
          "faltasSeguidas": {
            "type": "number",
            "example": 3
          },
          "faltasSeguidasComposicao": {
            "$ref": "#/components/schemas/FaltasSeguidasComposicaoResponseDto"
          },
          "turmaId": {
            "type": "string",
            "format": "uuid"
          },
          "turmaNome": {
            "type": "string",
            "nullable": true,
            "example": "Turma A"
          },
          "naTurmaHoje": {
            "type": "boolean"
          },
          "cobertura": {
            "$ref": "#/components/schemas/CoberturaResponseDto"
          }
        },
        "required": [
          "frequenciaPct",
          "confianca",
          "base",
          "presente",
          "ausente",
          "justificado",
          "faltasSeguidas",
          "faltasSeguidasComposicao",
          "turmaId",
          "turmaNome",
          "naTurmaHoje",
          "cobertura"
        ]
      },
      "OcorrenciaDoAlunoResponseDto": {
        "type": "object",
        "properties": {
          "turmaId": {
            "type": "string",
            "format": "uuid"
          },
          "turmaNome": {
            "type": "string",
            "nullable": true,
            "example": "Turma A"
          },
          "ocupacaoId": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "type": "string",
            "example": "2026-09-01"
          },
          "cancelada": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "presente",
              "ausente",
              "justificado"
            ]
          }
        },
        "required": [
          "turmaId",
          "turmaNome",
          "ocupacaoId",
          "data",
          "cancelada",
          "status"
        ]
      },
      "FrequenciaDoAlunoResponseDto": {
        "type": "object",
        "properties": {
          "alunoId": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "alunoAtivo": {
            "type": "boolean"
          },
          "vinculo": {
            "type": "string",
            "enum": [
              "pendente",
              "aprovado",
              "recusado"
            ]
          },
          "janelaDias": {
            "type": "number",
            "example": 30
          },
          "agregado": {
            "$ref": "#/components/schemas/AgregadoDeFrequenciaResponseDto"
          },
          "porTurma": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FrequenciaPorTurmaResponseDto"
            }
          },
          "ocorrencias": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcorrenciaDoAlunoResponseDto"
            }
          }
        },
        "required": [
          "alunoId",
          "nome",
          "alunoAtivo",
          "vinculo",
          "janelaDias",
          "agregado",
          "porTurma",
          "ocorrencias"
        ]
      },
      "AlunoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telefone": {
            "type": "string",
            "nullable": true,
            "example": "+5511999999999"
          },
          "nivelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo"
            ]
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "telefone",
          "nivelId",
          "status"
        ]
      },
      "AlunoPaginadoResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlunoResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137
          }
        },
        "required": [
          "data",
          "page",
          "pageSize",
          "total"
        ]
      },
      "AlunoComSenhaTemporariaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "telefone": {
            "type": "string",
            "nullable": true,
            "example": "+5511999999999"
          },
          "nivelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo"
            ]
          },
          "senhaTemporaria": {
            "type": "string",
            "example": "Xk4p-9Qm2"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "telefone",
          "nivelId",
          "status",
          "senhaTemporaria"
        ]
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "nivelId": {
            "type": "string"
          }
        },
        "required": [
          "nome",
          "email"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "nivelId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo"
            ]
          }
        }
      },
...
      "FotoDeProfessorResponseDto": {
        "type": "object",
        "properties": {
          "fotoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "fotoUrl"
        ]
      },
...
      "MatriculaEmTurmaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "turmaId": {
            "type": "string",
            "format": "uuid"
          },
          "alunoId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "turmaId",
          "alunoId",
          "createdAt"
        ]
      },
...
    }
  }
}